make chatElicitationRequestPart generic
